refactor(scripts): split consumer-compat plan, fixture, and CLI
consumer-compat.ts mixed pure package-manager planning, SvelteKit fixture
templates, and the pack/install harness. Extract plan and fixture modules,
share plot/package-directory constants, and register the new files on the
consumer CI lane and content-hash surface.
